tests: Move Eloquent Model tests into Repository
